requirements.txt
pandas==1.3.5
numpy==1.20.2
streamlit==1.16.0
plotly==5.6.0

requests==2.27.1

Pillow==8.2.0